Extract view model for Jetpack setup web view
