+++ WooCommerce/Classes/Authentication/Navigation Exceptions/PluginSetupWebViewController.swift
@@ -1,0 +1,106 @@
+import Combine
+import UIKit
+import WebKit
+
+/// The web view to handle plugin setup in the login flow.
+///
+final class PluginSetupWebViewController: UIViewController {
+
+    private let viewModel: PluginSetupWebViewModel
+
+    /// Main web view
+    private lazy var webView: WKWebView = {
+        let webView = WKWebView(frame: .zero)
+        webView.translatesAutoresizingMaskIntoConstraints = false
+        webView.navigationDelegate = self
+        return webView
+    }()
+
+    /// Progress bar for the web view
+    private lazy var progressBar: UIProgressView = {
+        let bar = UIProgressView(progressViewStyle: .bar)
+        bar.translatesAutoresizingMaskIntoConstraints = false
+        return bar
+    }()
+
+    /// Strong reference for the subscription to update progress bar
+    private var progressSubscription: AnyCancellable?
+
+    init(viewModel: PluginSetupWebViewModel) {
+        self.viewModel = viewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        configureNavigationBar()
+        configureWebView()
+        configureProgressBar()
+        startLoading()
+    }
+
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        if isMovingFromParent {
+            viewModel.trackDismissal()
+        }
+    }
+}
+
+private extension PluginSetupWebViewController {
+    func configureNavigationBar() {
+        title = viewModel.title
+    }
+
+    func configureWebView() {
+        view.addSubview(webView)
+        NSLayoutConstraint.activate([
+            view.leadingAnchor.constraint(equalTo: webView.leadingAnchor),
+            view.trailingAnchor.constraint(equalTo: webView.trailingAnchor),
+            view.safeTopAnchor.constraint(equalTo: webView.topAnchor),
+            view.bottomAnchor.constraint(equalTo: webView.bottomAnchor),
+        ])
+    }
+
+    func configureProgressBar() {
+        view.addSubview(progressBar)
+        NSLayoutConstraint.activate([
+            view.leadingAnchor.constraint(equalTo: progressBar.leadingAnchor),
+            view.trailingAnchor.constraint(equalTo: progressBar.trailingAnchor),
+            view.safeTopAnchor.constraint(equalTo: progressBar.topAnchor)
+        ])
+    }
+
+    func startLoading() {
+        guard let url = viewModel.initialURL else {
+            return
+        }
+        progressSubscription = webView.publisher(for: \.estimatedProgress)
+            .sink { [weak self] progress in
+                if progress == 1 {
+                    self?.progressBar.setProgress(0, animated: false)
+                } else {
+                    self?.progressBar.setProgress(Float(progress), animated: true)
+                }
+            }
+        let request = URLRequest(url: url)
+        webView.load(request)
+    }
+}
+
+extension PluginSetupWebViewController: WKNavigationDelegate {
+    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        guard let navigationURL = navigationAction.request.url else {
+            return
+        }
+        viewModel.decidePolicy(for: navigationURL, decisionHandler: decisionHandler)
+    }
+
+    func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
+        progressBar.setProgress(0, animated: false)
+    }
+}
